refactoring code based on feedback
