+++ src/main/java/ba/edu/ibu/bookreviewapp/core/api/CategorySuggestor.java
@@ -1,4 +1,4 @@
-package ba.edu.ibu.bookreviewapp.core.api.categorysuggestor;
+package ba.edu.ibu.bookreviewapp.core.api;
 
 public interface CategorySuggestor {
     String suggestCategory(String title);
