+++ src/main/java/ba/edu/ibu/bookreviewapp/core/model/Book.java
@@ -1,7 +1,5 @@
 package ba.edu.ibu.bookreviewapp.core.model;
 
-import com.fasterxml.jackson.annotation.JsonBackReference;
-import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonManagedReference;
 import jakarta.persistence.*;
