+++ src/main/java/ba/edu/ibu/bookreviewapp/core/model/Category.java
@@ -2,7 +2,6 @@ package ba.edu.ibu.bookreviewapp.core.model;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import com.fasterxml.jackson.annotation.JsonManagedReference;
 import jakarta.persistence.*;
 import java.util.List;
