+++ src/main/java/ba/edu/ibu/bookreviewapp/core/repository/UserRepository.java
@@ -8,5 +8,4 @@ import org.springframework.stereotype.Repository;
 public interface UserRepository extends JpaRepository<User, Long> {
     // Find a user by email
     User findByEmail(String email);
-
 }
