+++ src/main/java/ba/edu/ibu/bookreviewapp/core/service/UserBookService.java
@@ -96,12 +96,4 @@ public class UserBookService {
 
         return savedReview;
     }
-
-
-
-
-
-
-
-
 }
