+++ src/main/java/ba/edu/ibu/bookreviewapp/rest/configuration/WebConfig.java
@@ -1,4 +1,4 @@
-package ba.edu.ibu.bookreviewapp.config;
+package ba.edu.ibu.bookreviewapp.rest.configuration;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
