+++ src/main/java/ba/edu/ibu/bookreviewapp/rest/controller/UserBookController.java
@@ -76,6 +76,4 @@ public class UserBookController {
         UserBook updatedReview = userBookService.updateReview(userBook);
         return ResponseEntity.ok(updatedReview);
     }
-
-
 }
